Ref#4 Implementation of VODownloader
VODownloader now is only responsible for downloading the scn file which
accpets url directly.
And callback in closure returning the VirtualObject on download complete.
When download complete, the ARVC move the downloaded VirtualObject
above the QR code.

Glitch to be fixed: Unable to download textures of the scn file now.

PS: Some code change by lint autocorrect can be ignored,
and the old VC is deleted in this commit